remove duplicated code, put it in small pkg mtl_utils
