+++ setup.py
@@ -1,0 +1,4 @@
+from setuptools import setup
+
+
+setup(name='mtl_utils', packages=['mtl_utils'])
